requirements.txt
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.117.0
groq==0.4.1
mistralai==0.0.12
python-dotenv==1.0.1
pyyaml==6.0.1